% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acinar.data}
\alias{acinar.data}
\title{parameter files estimated from acinar.data using splatEstimate}
\format{
parameters estimated using splatEstimate
}
\usage{
acinar.data
}
\description{
parameter files estimated from acinar.data using splatEstimate
}
\keyword{datasets}
