% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SplatParams-methods.R
\name{newParams}
\alias{newParams}
\alias{newSplatParams}
\title{New Params}
\usage{
newSplatParams(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link{setParams}}.}
}
\value{
New Params object.
}
\description{
Create a new Params object. Functions exist for each of the different
Params subtypes.
}
