% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_acinar}
\alias{params_acinar}
\title{A data frame with 1000 genes and 80 cells}
\format{
A data frame with 1000 genes and 80 cells
}
\usage{
params_acinar
}
\description{
A data frame with 1000 genes and 80 cells
}
\keyword{datasets}
