% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splat-simulate.R
\name{SCRIPsimBCVMeans}
\alias{SCRIPsimBCVMeans}
\title{Simulate BCV means}
\usage{
SCRIPsimBCVMeans(sim, params)
}
\arguments{
\item{data}{data are used to fit the mean-BCV trend for simulation.}

\item{sim}{SingleCellExperiment to add BCV means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated BCV means.
}
\description{
Simulate means for each gene in each cell that are adjusted to follow a
mean-variance trend using Biological Coefficient of Variation (BCV). Five models were available in SCRIP including common BCV models (GP-commonBCV, BGP-commonBCV), trended BCV models (GP-trendedBCV, BGP-trendedBCV) and model without BCV effect (BP).
}
