% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimGroupCellMeans}
\alias{SCRIPsimGroupCellMeans}
\title{Simulate Group CellMeans}
\usage{
SCRIPsimGroupCellMeans(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add cell means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with added cell means.
}
\description{
Simulate group cell means
}
