% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimLibSizes}
\alias{SCRIPsimLibSizes}
\title{Simulate library sizes}
\usage{
SCRIPsimLibSizes(sim, params, libsize)
}
\arguments{
\item{sim}{SingleCellExperiment to add library size to.}

\item{params}{SplatParams object with simulation parameters.}

\item{libsize}{Provide the library size directly instread of using parameters to estimate}
}
\value{
SingleCellExperiment with simulated library sizes.
}
\description{
Simulate expected library sizes. Typically a log-normal distribution is used but there is also the option to use a normal distribution. In this case any negative values are set to half the minimum non-zero value.
}
