% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimPathDE}
\alias{SCRIPsimPathDE}
\title{Sim PathDE}
\usage{
SCRIPsimPathDE(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add dropout to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with DE for path simulation.
}
\description{
simulate DE factors for path
}
