% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimSingleCellMeans}
\alias{SCRIPsimSingleCellMeans}
\title{Simulate cell means}
\usage{
SCRIPsimSingleCellMeans(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add cell means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with added cell means.
}
\description{
Simulate a gene by cell matrix giving the mean expression for each gene in
each cell. Cells start with the mean expression for the group they belong to
(when simulating groups) or cells are assigned the mean expression from a
random position on the appropriate path (when simulating paths). The selected
means are adjusted for each cell's expected library size.
}
