% Generated by roxygen2: do not edit by hand
\name{simu.VEGs}
\alias{simu.VEGs}
\title{Simulation for clustering analysis using SCRIP (Application)}
\usage{
simu.VEGs(
  counts.matrix,
  params,
  base_allcellmeans,
  mode="GP-trendedBCV",
  nCells,
  nfeatures=1000
)
}
\arguments{
\item{counts.matrix}{data matrix required for simulation.}

\item{params}{SplatParams object containing parameters for the simulation.}

\item{base_allcellmeans}{base cell means specified directly for simulating counts.}

\item{mode}{"GP-commonBCV", "BP-commonBCV", "BP", "BGP-commonBCV" and "BGP-trendedBCV".}

\item{ncells}{number of cells simulated.}

\item{nfeatures}{parameter required for FinalVariable function in Seurat package .}
}
\value{
counts matrix
}
\description{
SCRIP simulation for clustering analysis
}
\details{
This function shows how to apply SCRIP for evaluating clustering analysis by preserving variably expressed genes.
}


