% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimDE.R
\name{simu_DE}
\alias{simu_DE}
\title{#SCRIP group simulation for differential expression analysis}
\usage{
simu_DE(
  expre_data,
  params,
  ngenes = NULL,
  nDE,
  ncells = NULL,
  FC,
  Dropout_rate = NULL,
  libsize = NULL,
  pre.bcv.df = NULL,
  bcv.shrink = 1,
  seed = 2021
)
}
\description{
#SCRIP group simulation for differential expression analysis
}
