% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimPathCellMeans}
\alias{SCRIPsimPathCellMeans}
\title{sim PathCellMeans}
\usage{
SCRIPsimPathCellMeans(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add dropout to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with cell means for path simulation.
}
\description{
simulate cell means for path
}
