% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimBatchEffects}
\alias{SCRIPsimBatchEffects}
\title{Simulate batch effects}
\usage{
SCRIPsimBatchEffects(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add batch effects to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated batch effects.
}
\description{
Simulate batch effects. Batch effect factors for each batch are produced
using \code{\link{getLNormFactors}} and these are added along with updated
means for each batch.
}
