% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimu.R
\name{SCRIPsimBCVMeans}
\alias{SCRIPsimBCVMeans}
\title{Simulate BCV means}
\usage{
SCRIPsimBCVMeans(data, sim, params)
}
\arguments{
\item{data}{data are used to fit the mean-BCV trend for simulation}

\item{sim}{SingleCellExperiment to add BCV means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated BCV means.
}
\description{
Simulate means for each gene in each cell that are adjusted to follow a
mean-variance trend using Biological Coefficient of Variation taken from
and inverse gamma distribution.
}
