% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimCluster.R
\name{simu.VEGs}
\alias{simu.VEGs}
\title{SCRIP simulation for clustering analysis}
\usage{
simu.VEGs(
  counts.matrix,
  params = params,
  base_allcellmeans,
  mode = "GP-trendedBCV",
  nCells,
  nfeatures = 1000
)
}
\arguments{
\item{counts.matrix}{data matrix required for simulation}

\item{params}{SplatParams object containing parameters for the simulation}

\item{base_allcellmeans}{base cell means specified directly for simulating counts}

\item{mode}{"GP-commonBCV", "BP-commonBCV", "BP", "BGP-commonBCV" and "BGP-trendedBCV"}

\item{nCells}{number of cells simulated}

\item{nfeatures}{parameter required for FinalVariable function in Seurat package}
}
\value{
simulated read counts data
}
\description{
Simulate count data for clustering analysis by preserving variably expressed genes
}
