% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRIPsimCluster.R
\name{simu.VEGs}
\alias{simu.VEGs}
\title{#SCRIP group simulation function for clustering analysis}
\usage{
simu.VEGs(
  counts.matrix,
  params = params,
  base_allcellmeans,
  mode = "GP-trendedBCV",
  nCells,
  nfeatures = 1000
)
}
\description{
#SCRIP group simulation function for clustering analysis
}
