% Generated by roxygen2: do not edit by hand
\name{simu_DE}
\alias{simu_DE}
\title{Simulation for differential expression analysis using SCRIP (Application)}
\usage{
simu_DE(
  expre_data,
  params,
  ngenes=NULL,
  nDE,
  ncells=NULL,
  FC,
  Dropout_rate=NULL,
  libsize=NULL,
  pre.bcv.df=NULL,
  bcv.shrink=1,
  seed=2021
)
}
\arguments{
\item{expre_data}{data matrix required for simulation.}

\item{params}{SplatParams object containing parameters for the simulation.}

\item{ngenes}{number of genes simulated.}

\item{nDE}{number of differentially expressed genes simulated.}

\item{ncells}{number of cells simulated.}

\item{FC}{fold change rate simulated between two groups.}

\item{Dropout_rate}{factor to control the dropout rate directly.}

\item{libsize}{library size used for simulation.}

\item{pre.bcv.df}{BCV.df enables us to change the variation of BCV values.}

\item{bcv.shrink}{seed for simulation.}
}
\value{
SingleCellExperiment containing simulated counts
}
\description{
Simulate counts for differential expression analysis using SCRIP method.
}
\details{
This function shows how to apply SCRIP for evaluating differential expression analysis methods. SCRIP enabled to simulate with different parameters including DE rates, fold change, dropout rate, library size, BCV.df and BCV.shrink.
}


